Split iOS unit and integration tests
