+++ native/ios-app/project.yml
@@ -25,6 +25,18 @@ schemes:
         - name: FireTests
           parallelizable: true
           randomExecutionOrder: true
+  FireIntegrationTests:
+    build:
+      targets:
+        Fire: all
+        FireIntegrationTests:
+          - test
+    test:
+      gatherCoverageData: false
+      targets:
+        - name: FireIntegrationTests
+          parallelizable: true
+          randomExecutionOrder: true
   FireUnitTests:
     build:
       targets:
@@ -126,6 +138,20 @@ targets:
     deploymentTarget: "17.0"
     sources:
       - path: Tests/Unit
+    dependencies:
+      - target: Fire
+    settings:
+      base:
+        GENERATE_INFOPLIST_FILE: YES
+        SWIFT_INCLUDE_PATHS:
+          - "$(inherited)"
+          - "$(SRCROOT)/Generated"
+  FireIntegrationTests:
+    type: bundle.unit-test
+    platform: iOS
+    deploymentTarget: "17.0"
+    sources:
+      - path: Tests/Integration
     dependencies:
       - target: Fire
     settings:
